Moved data fetching logic to smart component and moved categories component to component folder
